Move incomes and income statements out of PersonContext
